refactor: clarify responsibilities by restructuring OpenAI Gateway and renaming classes
